tp: refactor TrackEventParser;;EventImporter into separate class
Makes it possible to split this class up into legacy/non-legacy parsing.
